fix: restart the server with new options on configuration change
Currently, when users change their vscode configuration (say when they enable
Ivy mode), they'll have to manually reload the window to force the new
configuration to take effect.

However, we cannot just start and stop the `LanguageClient` connection
because the server options passed to the client initally is immutable.

To fix this, we wrap the `LanguageClient` in our own class
`AngularLanguageClient`, and reconstruct the server options every time the
server is restarted.

Code in `client/src/extension.ts` is copied to `client/src/client.ts` to
avoid circular dependency between `extension.ts` and `commands.ts`.
